requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
distlib==0.3.8
dnspython==2.6.1
email_validator==2.2.0
filelock==3.16.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn
idna==3.8
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.1.1
ordered-set==4.1.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.2
Pygments==2.18.0
rich==13.8.1
SQLAlchemy==2.0.34
typing_extensions==4.12.2
virtualenv==20.26.4
Werkzeug==3.0.4
wrapt==1.16.0
WTForms==3.1.2